Stop logging to STDERR by default
We currently log to `$stderr`. This changes it to log to `$stdout` but
only at the `FATAL` level by default. If `DEBUG` is set as an env var
then it will log at the `DEBUG` level.

I added `logger` to the Gemfile to address this Ruby 3.4 warning:

```
/home/runner/work/delivery_boy/delivery_boy/lib/delivery_boy.rb:1: warning: logger was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
You can add logger to your Gemfile or gemspec to silence this warning.
```